Rename orionbelt converter package to ossie_orionbelt for Apache naming consistency
Align with the OSI->Ossie incubation rename applied to the other converters:
package dir osi_orionbelt -> ossie_orionbelt, dist name apache-ossie-orionbelt,
CLI ossie-orionbelt. Keeps osi in spec-format file/class names (osi_to_obml.py,
OSItoOBML) as upstream did for dbt/gooddata.
