+++ converters/orionbelt/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "osi-orionbelt"
+name = "apache-ossie-orionbelt"
 version = "0.1.0"
 description = "Bidirectional OBML <-> OSI (Open Semantic Interchange) converter for OrionBelt semantic models"
 license = { text = "Apache-2.0" }
@@ -31,16 +31,16 @@ dev = [
 ]
 
 [project.scripts]
-osi-orionbelt = "osi_orionbelt.cli:main"
+ossie-orionbelt = "ossie_orionbelt.cli:main"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/osi_orionbelt"]
+packages = ["src/ossie_orionbelt"]
 
 # All three schemas (the two OSI artefacts plus a vendored snapshot of the
-# canonical OBML schema) live in src/osi_orionbelt/schemas/ as tracked files,
+# canonical OBML schema) live in src/ossie_orionbelt/schemas/ as tracked files,
 # so the sdist and wheel are self-contained and build in isolation. The OBML
 # snapshot is kept in sync with the repo-root schema/obml-schema.json by a
-# drift-guard test (tests/unit/test_osi_orionbelt_schema_sync.py).
+# drift-guard test (tests/unit/test_ossie_orionbelt_schema_sync.py).
 
 [tool.ruff]
 line-length = 100
@@ -51,7 +51,7 @@ select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]
 
 [tool.mypy]
 python_version = "3.12"
-files = ["src/osi_orionbelt"]
+files = ["src/ossie_orionbelt"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
