+++ converters/orionbelt/tests/fixtures/obml_as_osi.yaml
@@ -784,4 +784,4 @@ semantic_model:
     description: Average Sale
   custom_extensions:
   - vendor_name: COMMON
-    data: '{"source_format": "OBML", "source_version": "1.0", "converter": "osi-orionbelt"}'
+    data: '{"source_format": "OBML", "source_version": "1.0", "converter": "ossie-orionbelt"}'
