+++ converters/orionbelt/tests/test_osi_converter_filters.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 import json
 from typing import Any
 
-import osi_orionbelt.converter as conv
+import ossie_orionbelt.converter as conv
 
 _OBML_WITH_FILTERS: dict[str, Any] = {
     "version": 1.0,
