+++ converters/orionbelt/tests/test_osi_converter_measure_overrides.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from typing import Any
 
-import osi_orionbelt.converter as conv
+import ossie_orionbelt.converter as conv
 
 _BASE_OBML: dict[str, Any] = {
     "version": 1.0,
