+++ converters/orionbelt/tests/test_osi_converter_properties.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from typing import Any
 
-import osi_orionbelt.converter as conv
+import ossie_orionbelt.converter as conv
 
 _OBML_FULL: dict[str, Any] = {
     "version": 1.0,
